requirements.txt
pycrypto
fuzzywuzzy
